Cart and cart item adding, removing

+++ store/models.py
@@ -8,7 +8,7 @@ class Product(models.Model):
     title = models.CharField(max_length=255, unique=True)
     slug = models.SlugField(max_length=255, unique=True)
     description = models.TextField(max_length=500, blank=True)
-    price = models.DecimalField(max_digits=100, decimal_places=2)
+    price = models.FloatField()
     image = models.ImageField(upload_to="photos/products")
     stock = models.IntegerField()
     is_available = models.BooleanField(default=True)
